+++ packages/annotation-generator/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @sap-ux/annotation-generator
 
+## 0.3.22
+
+### Patch Changes
+
+-   d873cbd: chore - Add repository metadata in package.json
+-   Updated dependencies [d873cbd]
+    -   @sap-ux/fiori-annotation-api@0.4.22
+
+## 0.3.21
+
+### Patch Changes
+
+-   Updated dependencies [c3ebc82]
+    -   @sap-ux/project-access@1.29.18
+    -   @sap-ux/fiori-annotation-api@0.4.21
+
 ## 0.3.20
 
 ### Patch Changes