+++ packages/ui5-library-reference-writer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/ui5-library-reference-writer
 
+## 0.1.57
+
+### Patch Changes
+
+-   Updated dependencies [c3ebc82]
+    -   @sap-ux/project-access@1.29.18
+
 ## 0.1.56
 
 ### Patch Changes